+++ trixnity-messenger/src/webMain/kotlin/de/connect2x/trixnity/messenger/CreateRepositoriesModule.web.kt
@@ -1,9 +1,23 @@
 package de.connect2x.trixnity.messenger
 
 import de.connect2x.trixnity.messenger.util.RootPath
+import okio.FileSystem
 import org.koin.core.module.Module
 import org.koin.dsl.module
 
 actual fun platformCreateRepositoriesModuleModule(): Module = module {
-    single<CreateRepositoriesModule> { IndexedDBCreateRepositoriesModule(rootPath = get<RootPath>()) }
+    single<CreateRepositoriesModule> {
+        val matrixMessengerConfiguration = get<MatrixMessengerConfiguration>()
+
+        if (matrixMessengerConfiguration.features.enableRoom3ForNewAccountsOnWeb) {
+            AdaptiveCreateRepositoriesModule(
+                rootPath = get<RootPath>(),
+                fileSystem = getOrNull<FileSystem>(),
+                matrixMessengerConfiguration = matrixMessengerConfiguration,
+                scope = this@single,
+            )
+        } else {
+            IndexedDBCreateRepositoriesModule(rootPath = get<RootPath>())
+        }
+    }
 }